requirements.txt
streamlit==1.50.0
pandas==2.3.3
plotly==6.7.0
requests==2.32.5
requests-aws4auth==1.3.1
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
python-dotenv==1.2.1
